requirements.txt
adbutils==0.8.1
allure-pytest==2.8.16
allure-python-commons==2.8.16
apkutils2==1.0.0
Appium-Python-Client==1.0.1
atomicwrites==1.4.0
attrdict==2.0.1
attrs==19.3.0
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
cigam==0.0.3
colorama==0.4.3
decorator==4.4.2
Deprecated==1.2.10
deprecation==2.1.0
enum34==1.1.10
facebook-wda==0.9.6
idna==2.10
importlib-metadata==1.6.0
logzero==1.5.0
lxml==4.5.1
more-itertools==8.3.0
namedlist==1.7
numpy==1.19.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
progress==1.5
py==1.8.1
pyelftools==0.26
pyparsing==2.4.7
pytest==5.4.2
pytest-html==2.1.1
pytest-metadata==1.9.0
pytest-ordering==0.6
pytest-rerunfailures==9.0
requests==2.24.0
retry==0.9.2
selenium==3.141.0
six==1.15.0
tornado==6.0.4
uiautomator2==2.10.0
urllib3==1.25.9
wcwidth==0.1.9
weditor==0.5.0
whichcraft==0.6.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
